+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Vault Ruby Changelog
 
+## v0.13.0 (April 20, 2020)
+
+IMPROVEMENTS
+
+- Added support for defining a namespace when initializing the client, as well as options for changing the namespace via method.
+- Added support for sys/namespaces API. Ability to Get, Create, Delete, and List namespaces has been provided.
+
 ## v0.12.0 (August 14, 2018)
 
 IMPROVEMENTS